Fix type in av1_count_overlappable_neighbors()
mbmi->overlappable_neighbors is a `uint8_t`, but it was being
accessed through an `int *` in increment_int_ptr() (called by
av1_count_overlappable_neighbors()).

Fixing this type prevents issues on some rarer systems, such as
big-endian systems, or systems which expect `int`s to be aligned
to 4 bytes

+++ av1/common/reconinter.c
@@ -995,7 +995,7 @@ static INLINE void increment_int_ptr(MACROBLOCKD *xd, int rel_mi_row,
   (void)op_mi_size;
   (void)dir;
   (void)mi;
-  ++*(int *)fun_ctxt;
+  ++*(uint8_t *)fun_ctxt;
   (void)num_planes;
 }
